# go/docs/api/openapi-ru.yaml
openapi: 3.0.3
info:
  title: BizOps360 Go API
  description: |
    Полная спецификация API для платформы BizOps360 Go Backend.
    
    **Всего эндпоинтов:** 15
    
    **Базовая информация:**
    - Версия API: 1.0.0
    - Базовый URL (DEV): https://bizops360-api-go-dev-gqqr4r256q-uc.a.run.app
    - Формат данных: JSON
    - Кодировка: UTF-8
    
    **Аутентификация:**
    - Некоторые эндпоинты требуют API ключ в заголовке `X-Api-Key`
    - Эндпоинты здоровья, календаря, Zapier и V1 не требуют аутентификации
    
    **Ограничения:**
    - Rate limiting: 100 запросов в минуту на IP
    - Максимальный размер запроса: 1MB
  version: 1.0.0
  contact:
    name: BizOps360 Support
    url: https://github.com/bizops360/bizops360
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки (Local Development)
  - url: https://bizops360-api-go-dev-gqqr4r256q-uc.a.run.app
    description: Сервер разработки (Development)
  - url: https://bizops360-api-go-prod-XXXXX-uc.a.run.app
    description: Сервер продакшена (Production)
...
paths:
...
  /api/stripe/deposit:
    post:
      tags:
        - Stripe
      summary: Создание депозита
      description: Генерация инвойса для депозита бронирования
      operationId: createDeposit
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequest'
      responses:
        '200':
          description: Успешное создание депозита
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /api/stripe/deposit/calculate:
    get:
      tags:
        - Stripe
      summary: Расчет депозита
      description: Расчет рекомендуемого депозита на основе оценки стоимости
      operationId: calculateDeposit
      security:
        - ApiKeyAuth: []
      parameters:
        - name: estimate
          in: query
          required: true
          description: Оценка стоимости в долларах
          schema:
            type: number
            format: float
        - name: deposit
          in: query
          required: false
          description: Уже известный депозит (опционально)
          schema:
            type: number
            format: float
      responses:
        '200':
          description: Успешный расчет
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositCalculationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/stripe/deposit/with-email:
    post:
      tags:
        - Stripe
      summary: Создание депозита с отправкой email
      description: Генерация инвойса и отправка email (end-to-end процесс)
      operationId: createDepositWithEmail
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositWithEmailRequest'
      responses:
        '200':
          description: Успешное создание и отправка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositWithEmailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/stripe/final-invoice/with-email:
    post:
      tags:
        - Stripe
      summary: Создание финального инвойса с отправкой email (ORCHESTRATED)
      description: |
        **ОРКЕСТРИРОВАННЫЙ ЭНДПОИНТ** - Выполняет полный процесс:
        1. Создает финальный инвойс в Stripe для оставшегося баланса
        2. Автоматически отправляет кастомный email с деталями инвойса
        
        Это единая точка входа для создания и отправки финального инвойса.
        Используйте этот эндпоинт вместо последовательных вызовов `/api/stripe/final-invoice` и `/api/email/final-invoice`.
      operationId: createFinalInvoiceWithEmail
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalInvoiceRequest'
      responses:
        '200':
          description: Инвойс создан и email отправлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalInvoiceWithEmailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/stripe/test:
    post:
      tags:
        - Stripe
      summary: Тест интеграции Stripe
      description: Тестирование интеграции со Stripe
      operationId: testStripe
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Тест выполнен
          content:
            application/json:
              schema:
                type: object
  /api/estimate:
    post:
      tags:
        - Расчет стоимости
      summary: Расчет стоимости мероприятия
      description: Расчет стоимости мероприятия на основе даты, длительности и количества помощников
      operationId: calculateEstimate
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstimateRequest'
      responses:
        '200':
          description: Успешный расчет
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/estimate/special-dates:
    get:
      tags:
        - Расчет стоимости
      summary: Получить специальные даты
      description: Получить все специальные даты (праздники + даты повышенного спроса) на следующие N лет
      operationId: getSpecialDates
      security:
        - ApiKeyAuth: []
      parameters:
        - name: years
          in: query
          required: false
          description: Количество лет вперед (1-20, по умолчанию 5)
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
        - name: startYear
          in: query
          required: false
          description: Начальный год (2020-2100, по умолчанию текущий год)
          schema:
            type: integer
            minimum: 2020
            maximum: 2100
      responses:
        '200':
          description: Список специальных дат
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecialDatesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/email/test:
    post:
      tags:
        - Email
      summary: Отправка тестового email
      description: Отправка тестового email для проверки работы сервиса
      operationId: sendTestEmail
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestEmailRequest'
      responses:
        '200':
          description: Email отправлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/email/booking-deposit:
    post:
      tags:
        - Email
      summary: Отправка email о депозите бронирования
      description: Отправка email с информацией о депозите бронирования
      operationId: sendBookingDepositEmail
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingDepositEmailRequest'
      responses:
        '200':
          description: Email отправлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/calendar/create:
    post:
      tags:
        - Календарь
      summary: Создание события в календаре
      description: Создание события в Google Calendar (повторяет функциональность Apps Script createEvent)
      operationId: createCalendarEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalendarEventRequest'
      responses:
        '200':
          description: Событие создано успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarEventResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/business/{businessId}/process-lead:
    post:
      tags:
        - Обработка лидов
      summary: Обработка лида для бизнеса
      description: |
        Обработка лида для конкретного бизнеса. Выполняет полный цикл:
        1. Трансформация данных из Zapier
        2. Расчет стоимости
        3. Создание события в календаре
        4. Отправка email с квотацией
        5. Геокодирование адреса
      operationId: processBusinessLead
      parameters:
        - name: businessId
          in: path
          required: true
          description: Идентификатор бизнеса (например, stlpartyhelpers)
          schema:
            type: string
            example: stlpartyhelpers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZapierPayload'
      responses:
        '200':
          description: Лид успешно обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadProcessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Бизнес не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/zapier/process-lead:
    post:
      tags:
        - Обработка лидов
      summary: Обработка лида от Zapier (legacy)
      description: |
        Обработка лида от Zapier (legacy эндпоинт, повторяет Apps Script flow).
        Выполняет: расчет стоимости, отправку email с квотацией, создание события в календаре, геокодирование адреса.
      operationId: processZapierLead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZapierPayload'
      responses:
        '200':
          description: Лид успешно обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadProcessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/form-events:
    post:
      tags:
        - V1 Pipeline
      summary: Обработка события формы
      description: Обработка отправки формы (например, из WPForms)
      operationId: processFormEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormEventRequest'
      responses:
        '200':
          description: Событие обработано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormEventResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/triggers:
    post:
      tags:
        - V1 Pipeline
      summary: Обработка триггера
      description: Обработка триггерных событий (например, из Monday.com)
      operationId: processTrigger
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerRequest'
      responses:
        '200':
          description: Триггер обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API ключ для аутентификации
  schemas:
...
    DepositRequest:
      type: object
      required:
        - email
        - name
      properties:
        email:
          type: string
          format: email
          description: Email клиента
          example: "bizops-dev-alexey-at-shevelyov-dot-com@shevelyov.com"
        name:
          type: string
          description: Имя клиента
          example: "John Doe"
        estimatedTotal:
          type: number
          format: float
          description: Оценка общей стоимости
          example: 1000.0
        depositValue:
          type: number
          format: float
          nullable: true
          description: Значение депозита
          example: null
        deposit:
          type: number
          format: float
          nullable: true
          description: Депозит (альтернативное поле)
          example: null
        helpersCount:
          type: integer
          description: Количество помощников
          example: 2
        hours:
          type: number
          format: float
          description: Количество часов
          example: 4.0
        useTest:
          type: boolean
          default: false
          description: Использовать тестовый режим Stripe
          example: false
        dryRun:
          type: boolean
          default: false
          description: Тестовый запуск без реальных операций
          example: true
        mockStripe:
          type: boolean
          default: false
          description: Использовать мок Stripe
          example: false
      example:
        email: "bizops-dev-alexey-at-shevelyov-dot-com@shevelyov.com"
        name: "John Doe"
        estimatedTotal: 1000.0
        depositValue: null
        deposit: null
        helpersCount: 2
        hours: 4.0
        useTest: false
        dryRun: true
        mockStripe: false
    DepositResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        deposit:
          type: object
          properties:
            value:
              type: integer
              description: Значение в центах
            valueDollars:
              type: number
              format: float
              description: Значение в долларах
            percentage:
              type: number
              format: float
              description: Процент от общей стоимости
    DepositCalculationResponse:
      type: object
      properties:
        ok:
          type: boolean
        deposit:
          type: object
          properties:
            amountCents:
              type: integer
            amountDollars:
              type: number
              format: float
            percentage:
              type: number
              format: float
    DepositWithEmailRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "bizops-dev-alexey-at-shevelyov-dot-com@shevelyov.com"
        eventType:
          type: string
          example: "Birthday Party"
        eventDateTimeLocal:
          type: string
          example: "2025-06-15 17:00"
        eventDate:
          type: string
          format: date
          example: "2025-06-15"
        helpersCount:
          type: integer
          example: 2
        hours:
          type: number
          format: float
          example: 4.0
        duration:
          type: number
          format: float
          example: 4.0
        estimate:
          type: number
          format: float
          example: 1000.0
        estimatedTotal:
          type: number
          format: float
          example: 1000.0
        depositValue:
          type: number
          format: float
          nullable: true
          example: null
        useTest:
          type: boolean
          default: false
          example: false
        dryRun:
          type: boolean
          default: false
          example: true
        saveAsDraft:
          type: boolean
          default: false
          example: false
      example:
        name: "John Doe"
        email: "bizops-dev-alexey-at-shevelyov-dot-com@shevelyov.com"
        eventType: "Birthday Party"
        eventDateTimeLocal: "2025-06-15 17:00"
        eventDate: "2025-06-15"
        helpersCount: 2
        hours: 4.0
        duration: 4.0
        estimate: 1000.0
        estimatedTotal: 1000.0
        depositValue: null
        useTest: false
        dryRun: true
        saveAsDraft: false
    DepositWithEmailResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        invoiceId:
          type: string
        emailSent:
          type: boolean
    FinalInvoiceRequest:
      type: object
      required:
        - email
        - name
        - totalAmount
        - depositPaid
      properties:
        email:
          type: string
          format: email
          description: Email клиента
          example: "bizops-dev-alexey-at-shevelyov-dot-com@shevelyov.com"
        name:
          type: string
          description: Имя клиента
          example: "John Doe"
        totalAmountCents:
          type: integer
          description: Общая стоимость в центах
        totalAmount:
          type: number
          format: float
          description: Общая стоимость в долларах
          example: 1000.0
        depositPaidCents:
          type: integer
          description: Оплаченный депозит в центах
        depositPaid:
          type: number
          format: float
          description: Оплаченный депозит в долларах
          example: 400.0
        currency:
          type: string
          description: Валюта
          default: usd
          example: "usd"
        description:
          type: string
          description: Описание инвойса
        metadata:
          type: object
          additionalProperties:
            type: string
          description: Дополнительные метаданные
        useTest:
          type: boolean
          default: false
          description: Использовать тестовый режим Stripe
          example: false
      example:
        email: "bizops-dev-alexey-at-shevelyov-dot-com@shevelyov.com"
        name: "John Doe"
        totalAmount: 1000.0
        depositPaid: 400.0
        currency: "usd"
    FinalInvoiceWithEmailResponse:
      type: object
      properties:
        ok:
          type: boolean
          description: Общий успех операции
        message:
          type: string
          description: Сообщение о результате
          example: "Final invoice created and email sent"
        invoice:
          type: object
          properties:
            id:
              type: string
              description: ID инвойса в Stripe
            url:
              type: string
              format: uri
              description: URL для оплаты инвойса
            amount:
              type: number
              format: float
              description: Сумма к оплате в долларах
            status:
              type: string
              description: Статус инвойса
            pdf:
              type: string
              format: uri
              description: URL PDF инвойса
        details:
          type: object
          properties:
            totalAmount:
              type: number
              format: float
              description: Общая стоимость
            depositPaid:
              type: number
              format: float
              description: Оплаченный депозит
            remainingBalance:
              type: number
              format: float
              description: Оставшийся баланс
        email:
          type: object
          properties:
            sent:
              type: boolean
              description: Был ли отправлен email
            error:
              type: string
              nullable: true
              description: Ошибка при отправке email (если была)
      example:
        ok: true
        message: "Final invoice created and email sent"
        invoice:
          id: "in_1234567890"
          url: "https://invoice.stripe.com/i/acct_123/invoice_123"
          amount: 600.0
          status: "open"
        details:
          totalAmount: 1000.0
          depositPaid: 400.0
          remainingBalance: 600.0
        email:
          sent: true
          error: null
    EstimateRequest:
      type: object
      required:
        - eventDate
        - durationHours
        - numHelpers
      properties:
        eventDate:
          type: string
          format: date
          description: Дата мероприятия (YYYY-MM-DD)
          example: "2025-06-15"
        durationHours:
          type: number
          format: float
          description: Длительность в часах
          minimum: 0.1
          example: 4.0
        numHelpers:
          type: integer
          description: Количество помощников
          minimum: 1
          example: 2
      example:
        eventDate: "2025-06-15"
        durationHours: 4.0
        numHelpers: 2
...
    SpecialDatesResponse:
      type: object
      properties:
        ok:
          type: boolean
        yearsAhead:
          type: integer
          description: Количество лет вперед
        startYear:
          type: integer
          description: Начальный год
        data:
          type: object
          description: Объект с датами в формате YYYY-MM-DD
    TestEmailRequest:
      type: object
      required:
        - to
        - subject
      properties:
        to:
          type: string
          format: email
          description: Email получателя
          example: "bizops-dev-alexey-at-shevelyov-dot-com@shevelyov.com"
        subject:
          type: string
          description: Тема письма
          example: "Test Email"
        body:
          type: string
          description: Тело письма
          example: "<p>This is a test email</p>"
      example:
        to: "bizops-dev-alexey-at-shevelyov-dot-com@shevelyov.com"
        subject: "Test Email"
        body: "<p>This is a test email</p>"
    BookingDepositEmailRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          description: Имя клиента
        email:
          type: string
          format: email
          description: Email клиента
        eventDate:
          type: string
          description: Дата мероприятия
        estimatedTotal:
          type: number
          format: float
          description: Оценка общей стоимости
        depositAmount:
          type: number
          format: float
          description: Сумма депозита
    EmailResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Успешность отправки
        messageId:
          type: string
          nullable: true
          description: ID сообщения
        error:
          type: string
          nullable: true
          description: Ошибка при отправке
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Сообщение об ошибке
  responses:
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "invalid JSON: unexpected end of JSON input"
    Unauthorized:
      description: Требуется аутентификация
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "unauthorized: API key required"
    MethodNotAllowed:
      description: Метод не разрешен
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "method not allowed"